fix: Переделал jsx в tsx
